refactor: 별도 파일로 클래스 분리
