+++ src/main/kotlin/car/racing/CarRacing.kt
@@ -2,55 +2,6 @@ package car.racing
 
 import kotlin.random.Random
 
-class InputView(
-    private val reader: () -> String?,
-    private val console: (String) -> Unit
-) {
-
-    fun readDrivers(): Int {
-        console("자동차 대수는 몇 대인가요?: ")
-        return readInt()
-    }
-
-    fun readMoves(): Int {
-        console("시도할 횟수는 몇 회인가요?: ")
-        return readInt()
-    }
-
-    private fun readInt(): Int {
-        val value: String? = reader()
-        requireNotNull(value)
-        return kotlin
-            .runCatching { value.toInt() }
-            .getOrElse { throw IllegalArgumentException() }
-    }
-}
-
-class ResultView(
-    private val console: (String) -> Unit
-) {
-    fun showRecords(records: List<List<Int>>) {
-        console("실행 결과:\n")
-        records.forEach(::showRecord)
-    }
-
-    private fun showRecord(record: List<Int>) {
-        record.forEach { move ->
-            repeat(move) {
-                console(DISPLAY_MOVEMENT)
-            }
-            showEmptyLine()
-        }
-        showEmptyLine()
-    }
-
-    private fun showEmptyLine() = console("\n")
-
-    companion object {
-        private const val DISPLAY_MOVEMENT = "-"
-    }
-}
-
 class Car(
     private val forwardThreshold: Int = CAN_GO_FORWARD_THRESHOLD
 ) {
@@ -100,16 +51,4 @@ class CarRacing {
         require(moves > 0)
         return Game(drivers, moves)
     }
-}
-
-fun main() {
-    val inputView = InputView(::readLine, ::print)
-    val drivers = inputView.readDrivers()
-    val moves = inputView.readMoves()
-
-    val game = CarRacing().newGame(drivers, moves)
-    val records = game.play()
-
-    val resultView = ResultView(::print)
-    resultView.showRecords(records)
 }
